requirements.txt
fake-useragent==1.5.1
ipykernel==6.29.4
ipython==8.24.0jedi==0.19.1
jupyter_client==8.6.1
pandas==2.2.2
psycopg2-binary==2.9.9
pyarrow==16.1.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
urllib3==2.2.1